requirements.txt
boto3==1.4.4
botocore==1.5.17
python-docx==0.8.6
pdfrw==0.3
psycopg2==2.6.2
python-dotenv==0.6.3
